Rename drawbot modules
